requirements.txt
numpy>=1.26
pandas>=2.2
scipy>=1.11
scikit-learn>=1.4
PyYAML>=6.0
joblib>=1.3
torch>=2.2
gensim>=4.3
nltk>=3.8
